Update source again to make sure changes are persisted
Before, only the actual `source` object got updated, not the internal `_source`. The latter is used when exporting to an object through `toObject` though. So to make sure the changes are actually persisted, we need to call `updateSource` again.
If optimization is needed: calculate `totalStep` and `totalTarget` beforehand manually and only call `updateSource` once.