CSS and Accompanying EN Lang changes
This is the first large change to the css code. I had added my own code to the end of the css for now with only minimal changes made to the main body of the code (one or two minor ones that I didn't want to inheret and couldn't override).

I have also edited the EN lang file a little more, firstly so that some labels looked better on the sheet and remove superfluous words and secondly some I missed on the first pass or additions.