Implement leveled conditions by multiplying values
This is a first approach to handling status with multiple levels. If the value of the changes of the ActiveEffect are numeric, then multiply them by the level. This works for the current status with added changes, that is, Darkness, Cover, Overwhelmed, since they are just multiples of the base effect.