Basic Global Formatting + Unused .less moved to old-less folder
I have moved all files, declared unused by the earthdawn4e.less file, to a folder called old-less. This keeps them present but out of the way and helps me understand which I need to be working on.

I have made a few global changes to the .less file:

Window header formatting is complete.
A couple of global changes made, such as removal of backdrop-filters on actors from core css.
I have included the character sheet images and done some minor work to the less to prove the backgrounds function with the new headers.