sheet background less
Sets the PC, NPC and light backgrounds (added image files in webp for this)

Revamps the variables so they function with Light and Dark modes,.

Modifies the en.jason to add the SETTINGS.UI.FIELDS.colorScheme.x labels.

Fixes a slight padding error in the initial .less push that removed the padding at the bottom of windows.