Actor Sheet Re-Style
This is a big commit and involves .less and .hbs for character sheet styling. I have also made some changes to the en.json as I worked.

This is still something of a draft but I have made some alterations here that will knock on to sheets I have not refined: mostly going from partial__x to flex which will effect item sheets (these are still very draft so not a huge issue).