Actor Sheet Re-Style
This has added a little more interest to the light sheet (but it is still heavily a work in progress). This includes a mottled overlay on banner headings for both modes. This mottle must be .png as .webp was not working

Fixed a light-mode issue with the overwhelmed condition having a white background.

I have reordered some of the variable naming conventions to be more uniform.