Actor Sheet Re-Style
Recovery added to the attribute sidebar and the recovery test button added as an icon.

Footer abilities moved to the general window to remove the footer for now. I think the footer ought to be removed, however I can't do it as it requires .js, so I left it with a <div></div> in there for now so it all still works.

Some new en.json added to support this change.