Style Checkpoint
There have been a lot of changes made and this is a checkpoint to use as a base. Certain things need to be done which I can't, such as a removing superfluous sheets that are needed in the .js initialisation code when the sheet is opened (can't just be deleted). There is also no work done to the header, however the tabs are for the most part done in dark and need only minor visual changes to make light better functioning.