button-style-global
Ok this grew from the branch name somewhat:

buttons (fa awesome buttons) are moved to global, but I also redid the svg icons in line with the damage button (all functioning though equipment I need to ask about). This is more consistent in enabling styling.

This has also seen a HUGE overhaul of the variables, they are more bloated but it allows for more customisability of light and dark