language-container
Formatting complete on languages, but needs work with .js to finish.

All upgrades are not restricted to edit mode only and all sections are updated to contain the right class to do this in future.

all + - buffers are now on game mode only.

uuid icons are all formatted to appear like all other buttons (including hover over which was not working prior)